people-area-identify/requirements.txt
alibabacloud_dyvmsapi20170525==3.2.1
alibabacloud_tea_openapi==0.3.9
alibabacloud_tea_util==0.3.12
coremltools==7.2
matplotlib==3.8.4
numpy==1.26.4
onnx==1.16.1
opencv_python==4.8.1.78
pafy==0.5.5
pandas==2.2.2
Pillow==10.3.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.2
scipy==1.13.1
seaborn==0.13.2
setuptools==68.0.0
thop==0.1.1.post2209072238
torch==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
wandb==0.17.0
